# spec/openapi.json
{
  "components": {
    "schemas": {
      "AvailabilityRequest": {
        "description": "`POST /v1/reservations/availability` body (real-time, deadline-budgeted).",
        "properties": {
          "datetime": {
            "format": "date-time",
            "title": "Datetime",
            "type": "string"
          },
          "party_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Party Size",
            "type": "integer"
          },
          "tenant": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Tenant",
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "party_size",
          "datetime"
        ],
        "title": "AvailabilityRequest",
        "type": "object"
      },
      "AvailabilityResult": {
        "description": "`data` for availability: the normalized success state + candidate slots.\n\n`state` is `available` (slots present) or `unavailable` (none). Non-definitive\noutcomes (timeout/auth_error/\u2026) never reach here \u2014 they are raised as errors.",
        "properties": {
          "slots": {
            "items": {
              "$ref": "#/components/schemas/AvailabilitySlot"
            },
            "title": "Slots",
            "type": "array"
          },
          "state": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "slots"
        ],
        "title": "AvailabilityResult",
        "type": "object"
      },
      "AvailabilitySlot": {
        "description": "One bookable slot returned by availability.",
        "properties": {
          "party_size": {
            "title": "Party Size",
            "type": "integer"
          },
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          }
        },
        "required": [
          "time",
          "party_size"
        ],
        "title": "AvailabilitySlot",
        "type": "object"
      },
      "BookingRequest": {
        "description": "`POST /v1/reservations/book` body (a consequential write \u2014 Rule 9 / review #5).\n\n`idempotency_key` makes the write retry-safe: a retried `book` with the same key\nmust NOT create a second booking. The mock derives a deterministic\nconfirmation_id from the key to demonstrate this at the contract level; the real\nbackend forwards the key to OpenTable's idempotency mechanism (verified against\nOpenTable docs before that backend is written \u2014 Rule 2).",
        "properties": {
          "datetime": {
            "format": "date-time",
            "title": "Datetime",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "party_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Party Size",
            "type": "integer"
          },
          "tenant": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Tenant",
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "name",
          "party_size",
          "datetime"
        ],
        "title": "BookingRequest",
        "type": "object"
      },
      "BookingResult": {
        "description": "`data` for book: an actual confirmation only.\n\n`state` is always `confirmed` and `confirmation_id` is always present. Any\nnon-confirmed booking outcome (slot gone, ambiguous, timeout, \u2026) is raised as an\nerror so this success shape can never carry a false/empty confirmation.",
        "properties": {
          "confirmation_id": {
            "title": "Confirmation Id",
            "type": "string"
          },
          "state": {
            "const": "confirmed",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "confirmation_id"
        ],
        "title": "BookingResult",
        "type": "object"
      },
      "CancelRequest": {
        "description": "`POST /v1/reservations/cancel` body (a consequential write \u2014 Rule 9 / review #5).",
        "properties": {
          "confirmation_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Confirmation Id",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "tenant": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Tenant",
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "confirmation_id"
        ],
        "title": "CancelRequest",
        "type": "object"
      },
      "CancelResult": {
        "description": "`data` for cancel: the confirmation id + `cancelled` state.",
        "properties": {
          "confirmation_id": {
            "title": "Confirmation Id",
            "type": "string"
          },
          "state": {
            "const": "cancelled",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "confirmation_id"
        ],
        "title": "CancelResult",
        "type": "object"
      },
      "Envelope_AvailabilityResult_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AvailabilityResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "ok",
          "source",
          "latency_ms",
          "mock",
          "request_id"
        ],
        "title": "Envelope[AvailabilityResult]",
        "type": "object"
      },
      "Envelope_BookingResult_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "ok",
          "source",
          "latency_ms",
          "mock",
          "request_id"
        ],
        "title": "Envelope[BookingResult]",
        "type": "object"
      },
      "Envelope_CancelResult_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CancelResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "ok",
          "source",
          "latency_ms",
          "mock",
          "request_id"
        ],
        "title": "Envelope[CancelResult]",
        "type": "object"
      },
      "Envelope_ModifyResult_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModifyResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "ok",
          "source",
          "latency_ms",
          "mock",
          "request_id"
        ],
        "title": "Envelope[ModifyResult]",
        "type": "object"
      },
      "ErrorInfo": {
        "description": "The structured `error` object (present only when `ok` is false).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorInfo",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Liveness payload \u2014 intentionally tiny and secret-free.",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ModifyRequest": {
        "description": "`POST /v1/reservations/modify` body (a consequential write \u2014 Rule 9 / review #5).",
        "properties": {
          "confirmation_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Confirmation Id",
            "type": "string"
          },
          "datetime": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datetime"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "party_size": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Size"
          },
          "tenant": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Tenant",
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "confirmation_id"
        ],
        "title": "ModifyRequest",
        "type": "object"
      },
      "ModifyResult": {
        "description": "`data` for modify: the confirmation id + `modified` state.",
        "properties": {
          "confirmation_id": {
            "title": "Confirmation Id",
            "type": "string"
          },
          "state": {
            "const": "modified",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "confirmation_id"
        ],
        "title": "ModifyResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionInfo": {
        "description": "Contract/version info for the internal API.",
        "properties": {
          "api_version": {
            "title": "Api Version",
            "type": "string"
          },
          "integrations": {
            "items": {
              "type": "string"
            },
            "title": "Integrations",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "api_version",
          "version",
          "integrations"
        ],
        "title": "VersionInfo",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Switchboard \u2014 the Integration Gateway. A standalone, internal-only HTTP API that owns third-party integrations behind one clean, versioned, token-gated contract and returns a uniform response envelope with a normalized result state. Reservations (OpenTable) is the first integration, mock-first.",
    "title": "Switchboard \u2014 Integration Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Alias of /healthz for the documented Postman/CI healthcheck.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness probe (alias)",
        "tags": [
          "system"
        ]
      }
    },
    "/healthz": {
      "get": {
        "description": "Return liveness. No auth, no secrets.",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "system"
        ]
      }
    },
    "/v1": {
      "get": {
        "description": "Return the internal API's name, version, and live integrations.",
        "operationId": "version_info_v1_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionInfo"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Contract/version info",
        "tags": [
          "system"
        ]
      }
    },
    "/v1/reservations/availability": {
      "post": {
        "description": "Check reservation availability (real-time; deadline-budgeted).",
        "operationId": "availability_v1_reservations_availability_post",
        "parameters": [
          {
            "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
            "in": "header",
            "name": "X-Deadline-Ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
              "title": "X-Deadline-Ms"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AvailabilityResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reservation Availability v1 (real-time)",
        "tags": [
          "reservations"
        ]
      }
    },
    "/v1/reservations/book": {
      "post": {
        "description": "Create a booking (consequential write; idempotent; race-safe \u2014 review #5).\n\nReturns `ok=true` ONLY on an actual confirmation. A slot-gone race surfaces as\n`unavailable`, an ambiguous result as `requires_human` \u2014 never a false success.",
        "operationId": "book_v1_reservations_book_post",
        "parameters": [
          {
            "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
            "in": "header",
            "name": "X-Deadline-Ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
              "title": "X-Deadline-Ms"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_BookingResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reservation Booking v1 (consequential write)",
        "tags": [
          "reservations"
        ]
      }
    },
    "/v1/reservations/cancel": {
      "post": {
        "description": "Cancel an existing booking (consequential write; idempotent).",
        "operationId": "cancel_v1_reservations_cancel_post",
        "parameters": [
          {
            "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
            "in": "header",
            "name": "X-Deadline-Ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
              "title": "X-Deadline-Ms"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_CancelResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a reservation (consequential write)",
        "tags": [
          "reservations"
        ]
      }
    },
    "/v1/reservations/modify": {
      "post": {
        "description": "Modify an existing booking (consequential write; idempotent).",
        "operationId": "modify_v1_reservations_modify_post",
        "parameters": [
          {
            "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
            "in": "header",
            "name": "X-Deadline-Ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller's hard deadline in milliseconds. The effective deadline is min(this, the per-endpoint budget). On exceed, the endpoint returns state=timeout.",
              "title": "X-Deadline-Ms"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ModifyResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Modify a reservation (consequential write)",
        "tags": [
          "reservations"
        ]
      }
    }
  }
}